@@ -12,33 +12,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cdk": "^5.2.3",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.3",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "6.0.2",
+    "@angular/cdk": "^6.0.2",
+    "@angular/common": "6.0.2",
+    "@angular/compiler": "6.0.2",
+    "@angular/core": "6.0.2",
+    "@angular/forms": "6.0.2",
+    "@angular/http": "6.0.2",
+    "@angular/material": "^6.0.2",
+    "@angular/platform-browser": "6.0.2",
+    "@angular/platform-browser-dynamic": "6.0.2",
+    "@angular/router": "6.0.2",
     "@fortawesome/fontawesome": "^1.1.4",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
     "angular-carousel": "^1.0.2",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "node-sass": "^4.5.2",
-    "rxjs": "^5.5.2",
+    "node-sass": "^4.9.0",
+    "rxjs": "^6.1.0",
     "sass-loader": "^6.0.5",
     "underscore": "^1.8.3",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.1",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular/cli": "6.0.3",
+    "@angular/compiler-cli": "6.0.2",
+    "@angular/language-service": "6.0.2",
     "@types/jasmine": "2.8.3",
     "@types/node": "~6.0.60",
     "codelyzer": "^4.0.1",
@@ -53,7 +53,8 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
-    "webpack": "^3.11.0"
+    "typescript": "2.7.2",
+    "webpack": "^4.4.0",
+    "@angular-devkit/build-angular": "~0.6.3"
   }
 }
