@@ -1,7 +1,6 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
 import { AppComponent } from './app.component';
 import { RouterModule } from '@angular/router';
 import { PagenotfoundComponent } from './shared/pages/pagenotfound/pagenotfound.component';
@@ -11,17 +10,13 @@ import { MobileModule } from 'app/mobile/mobile.module';
 import { SharedModule } from 'app/shared/shared.module';
 import { AppRoutingModule } from 'app/app-routing.module';
 import { CoreModule } from 'app/core/core.module';
-import {HttpClientModule} from '@angular/common/http';
-import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
-import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import { HttpClientModule } from '@angular/common/http';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MyAngularMaterialModule } from 'app/shared/angular-material/angularMaterial.module';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    PagenotfoundComponent,
-    HomeComponent
-  ],
+  declarations: [AppComponent, PagenotfoundComponent, HomeComponent],
   imports: [
     NgbModule.forRoot(),
     BrowserModule,
@@ -40,4 +35,4 @@ import { MyAngularMaterialModule } from 'app/shared/angular-material/angularMate
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}
