@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Http, Response } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
+import { Observable } from 'rxjs';
+
+
 
 import { MenuTopNav } from '../interfaces/header';
 import { SearchResponse } from 'app/shared/ResourceModels/SearchResponse';
