@@ -1,6 +1,6 @@
-import { HttpHeaders } from "@angular/common/http";
+import { HttpHeaders } from '@angular/common/http';
 
-export class HTTPHeader{
+export class HTTPHeader {
   httpOptions = {
     headers: new HttpHeaders({
       'Content-Type': 'application/json'
